updates, fixes

+++ System/HyperScan/examples/USB/DOOM/i_video.c
@@ -214,7 +214,7 @@ void I_InitGraphics (void)
     }
 
     /* Allocate screen to draw to */
-	I_VideoBuffer = (unsigned int*)Z_Malloc (SCREENWIDTH * SCREENHEIGHT, PU_STATIC, NULL);  // For DOOM to draw on
+	I_VideoBuffer = (unsigned short*)Z_Malloc (SCREENWIDTH * SCREENHEIGHT, PU_STATIC, NULL);  // For DOOM to draw on
 
 	screenvisible = true;
